Combine and sort tests
