--- test/geocel/MultiLevelGeoTest.hh
@@ -1,36 +1,0 @@
-//------------------------------- -*- C++ -*- -------------------------------//
-// Copyright Celeritas contributors: see top-level COPYRIGHT file for details
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file geocel/MultiLevelGeoTest.hh
-//---------------------------------------------------------------------------//
-#pragma once
-
-#include <string_view>
-
-#include "GenericGeoTestInterface.hh"
-
-namespace celeritas
-{
-namespace test
-{
-//---------------------------------------------------------------------------//
-/*!
- * Test the multi-level geometry.
- */
-class MultiLevelGeoTest
-{
-  public:
-    static std::string_view geometry_basename() { return "multi-level"; }
-
-    MultiLevelGeoTest(GenericGeoTestInterface* geo_test);
-    void test_accessors() const;
-    void test_trace() const;
-
-  private:
-    GenericGeoTestInterface* test_;
-};
-
-//---------------------------------------------------------------------------//
-}  // namespace test
-}  // namespace celeritas
