--- test/geocel/ZnenvGeoTest.hh
@@ -1,67 +1,0 @@
-//------------------------------- -*- C++ -*- -------------------------------//
-// Copyright Celeritas contributors: see top-level COPYRIGHT file for details
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file geocel/ZnenvGeoTest.hh
-//---------------------------------------------------------------------------//
-#pragma once
-
-#include <string_view>
-
-#include "GenericGeoTestInterface.hh"
-#include "TestMacros.hh"
-
-namespace celeritas
-{
-namespace test
-{
-//---------------------------------------------------------------------------//
-/*!
- * Test the transformed box geometry.
- */
-class ZnenvGeoTest
-{
-  public:
-    static std::string_view geometry_basename() { return "znenv"; }
-
-    //! Construct with a reference to the GoogleTest
-    ZnenvGeoTest(GenericGeoTestInterface* geo_test) : test_{geo_test} {}
-
-    void test_accessors() const;
-    inline void test_trace() const;
-
-  private:
-    GenericGeoTestInterface* test_;
-};
-
-//---------------------------------------------------------------------------//
-void ZnenvGeoTest::test_trace() const
-{
-    // NOTE: This tests the capability of the G4PVDivision conversion based on
-    // an ALICE component
-    static char const* const expected_mid_volumes[] = {
-        "World", "ZNENV", "ZNST", "ZNST",  "ZNST",  "ZNST", "ZNST",
-        "ZNST",  "ZNST",  "ZNST", "ZNST",  "ZNST",  "ZNST", "ZNST",
-        "ZNST",  "ZNST",  "ZNST", "ZNST",  "ZNST",  "ZNST", "ZNST",
-        "ZNST",  "ZNST",  "ZNST", "ZNENV", "World",
-    };
-    static real_type const expected_mid_distances[] = {
-        6.38, 0.1,  0.32, 0.32, 0.32, 0.32, 0.32, 0.32,  0.32,
-        0.32, 0.32, 0.32, 0.32, 0.32, 0.32, 0.32, 0.32,  0.32,
-        0.32, 0.32, 0.32, 0.32, 0.32, 0.32, 0.1,  46.38,
-    };
-    {
-        auto result = test_->track({-10, 0.0001, 0}, {1, 0, 0});
-        EXPECT_VEC_EQ(expected_mid_volumes, result.volumes);
-        EXPECT_VEC_SOFT_EQ(expected_mid_distances, result.distances);
-    }
-    {
-        auto result = test_->track({0.0001, -10, 0}, {0, 1, 0});
-        EXPECT_VEC_EQ(expected_mid_volumes, result.volumes);
-        EXPECT_VEC_SOFT_EQ(expected_mid_distances, result.distances);
-    }
-}
-
-//---------------------------------------------------------------------------//
-}  // namespace test
-}  // namespace celeritas
