+++ test/geocel/g4/GeantGeo.test.cc
@@ -4,6 +4,7 @@
 //---------------------------------------------------------------------------//
 //! \file geocel/g4/GeantGeo.test.cc
 //---------------------------------------------------------------------------//
+#include <regex>
 #include <string_view>
 #include <G4LogicalVolume.hh>
 
@@ -58,6 +59,43 @@ class GeantGeoTest : public GeantGeoTestBase
 };
 
 //---------------------------------------------------------------------------//
+using CmseTest = GenericGeoParameterizedTest<GeantGeoTest, CmseGeoTest>;
+
+TEST_F(CmseTest, trace)
+{
+    this->impl().test_trace();
+}
+
+TEST_F(CmseTest, imager)
+{
+    SafetyImager write_image{this->geometry()};
+
+    ImageInput inp;
+    inp.lower_left = from_cm({-550, 0, -4000});
+    inp.upper_right = from_cm({550, 0, 2000});
+    inp.rightward = {0.0, 0.0, 1.0};
+    inp.vertical_pixels = 8;
+
+    std::string prefix = "g4";
+
+    write_image(ImageParams{inp}, prefix + "-cmse.jsonl");
+}
+
+//---------------------------------------------------------------------------//
+using CmsEeBackDeeTest
+    = GenericGeoParameterizedTest<GeantGeoTest, CmsEeBackDeeGeoTest>;
+
+TEST_F(CmsEeBackDeeTest, accessors)
+{
+    this->impl().test_accessors();
+}
+
+TEST_F(CmsEeBackDeeTest, trace)
+{
+    this->impl().test_trace();
+}
+
+//---------------------------------------------------------------------------//
 using FourLevelsTest
     = GenericGeoParameterizedTest<GeantGeoTest, FourLevelsGeoTest>;
 
@@ -65,8 +103,6 @@ TEST_F(FourLevelsTest, accessors)
 {
     this->impl().test_accessors();
 }
-
-//---------------------------------------------------------------------------//
 
 TEST_F(FourLevelsTest, consecutive_compute)
 {
@@ -87,8 +123,6 @@ TEST_F(FourLevelsTest, consecutive_compute)
     geo = {from_cm({-9, -10, -10}), {1, 0, 0}};
     EXPECT_SOFT_EQ(4.0, to_cm(geo.find_safety()));
 }
-
-//---------------------------------------------------------------------------//
 
 TEST_F(FourLevelsTest, detailed_track)
 {
@@ -147,13 +181,11 @@ TEST_F(FourLevelsTest, detailed_track)
     }
 }
 
-//---------------------------------------------------------------------------//
-
 TEST_F(FourLevelsTest, reentrant_boundary)
 {
     auto geo = this->make_geo_track_view({15.5, 10, 10}, {-1, 0, 0});
     ASSERT_FALSE(geo.is_outside());
-    EXPECT_EQ(VolumeId{1}, geo.volume_id());
+    EXPECT_EQ("Shape1", this->volume_name(geo));
     EXPECT_FALSE(geo.is_on_boundary());
 
     // Check for surfaces: we should hit the outside of the sphere Shape2
@@ -164,13 +196,13 @@ TEST_F(FourLevelsTest, reentrant_boundary)
     EXPECT_TRUE(geo.is_on_boundary());
     geo.set_dir({0, 1, 0});
     EXPECT_TRUE(geo.is_on_boundary());
-    EXPECT_EQ(VolumeId{1}, geo.volume_id());
+    EXPECT_EQ("Shape1", this->volume_name(geo));
 
     // Move a bit internally, then scatter back toward the sphere
     next = geo.find_next_step(from_cm(10.0));
     EXPECT_SOFT_EQ(6, to_cm(next.distance));
     geo.set_dir({-1, 0, 0});
-    EXPECT_EQ(VolumeId{1}, geo.volume_id());
+    EXPECT_EQ("Shape1", this->volume_name(geo));
 
     // Move to the sphere boundary then scatter still into the sphere
     next = geo.find_next_step(from_cm(10.0));
@@ -181,7 +213,8 @@ TEST_F(FourLevelsTest, reentrant_boundary)
     geo.set_dir({0, -1, 0});
     EXPECT_TRUE(geo.is_on_boundary());
     geo.cross_boundary();
-    EXPECT_EQ(VolumeId{0}, geo.volume_id());
+    EXPECT_EQ("Shape2", this->volume_name(geo));
+
     EXPECT_TRUE(geo.is_on_boundary());
 
     // Travel nearly tangent to the right edge of the sphere, then scatter to
@@ -193,19 +226,16 @@ TEST_F(FourLevelsTest, reentrant_boundary)
     geo.set_dir({1, 0, 0});
     EXPECT_TRUE(geo.is_on_boundary());
     geo.cross_boundary();
-    EXPECT_EQ(VolumeId{1}, geo.volume_id());
+    EXPECT_EQ("Shape1", this->volume_name(geo));
+
     EXPECT_TRUE(geo.is_on_boundary());
     next = geo.find_next_step(from_cm(10.0));
 }
-
-//---------------------------------------------------------------------------//
 
 TEST_F(FourLevelsTest, trace)
 {
     this->impl().test_trace();
 }
-
-//---------------------------------------------------------------------------//
 
 TEST_F(FourLevelsTest, safety)
 {
@@ -256,8 +286,6 @@ TEST_F(FourLevelsTest, safety)
     EXPECT_VEC_SOFT_EQ(expected_lim_safeties, lim_safeties);
 }
 
-//---------------------------------------------------------------------------//
-
 TEST_F(FourLevelsTest, levels)
 {
     auto geo = this->make_geo_track_view({10.0, 10.0, 10.0}, {1, 0, 0});
@@ -283,230 +311,6 @@ TEST_F(FourLevelsTest, levels)
     geo.cross_boundary();
 
     EXPECT_EQ("[OUTSIDE]", this->all_volume_instance_names(geo));
-}
-
-//---------------------------------------------------------------------------//
-class SolidsTest
-    : public GenericGeoParameterizedTest<GeantGeoTest, SolidsGeoTest>
-{
-    SpanStringView expected_log_levels() const final
-    {
-        if (geant4_version < Version{11})
-            return {};
-
-        static std::string_view const levels[] = {"error"};
-        return make_span(levels);
-    }
-};
-
-//---------------------------------------------------------------------------//
-TEST_F(SolidsTest, output)
-{
-    GeoParamsOutput out(this->geometry());
-    EXPECT_EQ("geometry", out.label());
-
-    if (CELERITAS_UNITS == CELERITAS_UNITS_CGS)
-    {
-        EXPECT_JSON_EQ(
-            R"json({"_category":"internal","_label":"geometry","bbox":[[-600.0,-300.0,-75.0],[600.0,300.0,75.0]],"max_depth":2,"supports_safety":true,"volumes":{"label":["","","","","box500","cone1","para1","sphere1","parabol1","trap1","trd1","trd2","","trd3_refl@1","tube100","boolean1","polycone1","genPocone1","ellipsoid1","tetrah1","orb1","polyhedr1","hype1","elltube1","ellcone1","arb8b","arb8a","xtru1","World","trd3_refl@0"]}})json",
-            to_string(out));
-    }
-}
-//---------------------------------------------------------------------------//
-
-TEST_F(SolidsTest, accessors)
-{
-    TestImpl(this).test_accessors();
-}
-
-//---------------------------------------------------------------------------//
-
-TEST_F(SolidsTest, trace)
-{
-    TestImpl(this).test_trace();
-}
-
-//---------------------------------------------------------------------------//
-
-TEST_F(SolidsTest, reflected_vol)
-{
-    auto geo = this->make_geo_track_view({-500, -125, 0}, {0, 1, 0});
-    EXPECT_EQ(VolumeId{29}, geo.volume_id());
-    auto const& label = this->geometry()->volumes().at(geo.volume_id());
-    EXPECT_EQ("trd3_refl", label.name);
-    EXPECT_FALSE(ends_with(label.ext, "_refl"));
-}
-
-TEST_F(SolidsTest, DISABLED_imager)
-{
-    SafetyImager write_image{this->geometry()};
-
-    ImageInput inp;
-    inp.lower_left = from_cm({-550, -250, 5});
-    inp.upper_right = from_cm({550, 250, 5});
-    inp.rightward = {1.0, 0.0, 0.0};
-    inp.vertical_pixels = 8;
-
-    write_image(ImageParams{inp}, "g4-solids-xy-hi.jsonl");
-
-    inp.lower_left[2] = inp.upper_right[2] = from_cm(-5);
-    write_image(ImageParams{inp}, "g4-solids-xy-lo.jsonl");
-}
-
-//---------------------------------------------------------------------------//
-// TODO: reorder or independently run tests; for now this has to go after the
-// Solids test
-using ReplicaTest = GenericGeoParameterizedTest<GeantGeoTest, ReplicaGeoTest>;
-
-TEST_F(ReplicaTest, trace)
-{
-    this->impl().test_trace();
-}
-
-TEST_F(ReplicaTest, volume_stack)
-{
-    this->impl().test_volume_stack();
-}
-
-TEST_F(ReplicaTest, level_strings)
-{
-    using R2 = Array<double, 2>;
-
-    auto const& geo_params = *this->geometry();
-    auto const& vol_inst = geo_params.volume_instances();
-
-    static R2 const points[] = {
-        {-435, 550},
-        {-460, 550},
-        {-400, 650},
-        {-450, 650},
-        {-450, 700},
-    };
-
-    std::vector<std::string> all_vol_inst;
-    for (R2 xz : points)
-    {
-        auto geo = this->make_geo_track_view({xz[0], 0.0, xz[1]}, {1, 0, 0});
-
-        auto level = geo.level();
-        CELER_ASSERT(level && level >= LevelId{0});
-        std::vector<VolumeInstanceId> inst_ids(level.get() + 1);
-        geo.volume_instance_id(make_span(inst_ids));
-        std::vector<std::string> names(inst_ids.size());
-        for (auto i : range(inst_ids.size()))
-        {
-            Label lab = vol_inst.at(inst_ids[i]);
-            if (auto phys_inst = geo_params.id_to_geant(inst_ids[i]))
-            {
-                if (phys_inst.replica)
-                {
-                    lab.ext += '+';
-                    lab.ext += std::to_string(phys_inst.replica.get());
-                }
-            }
-            names[i] = to_string(lab);
-        }
-        all_vol_inst.push_back(to_string(repr(names)));
-    }
-
-    static char const* const expected_all_vol_inst[] = {
-        R"({"world_PV", "fSecondArmPhys", "EMcalorimeter", "cell_param@+14"})",
-        R"({"world_PV", "fSecondArmPhys", "EMcalorimeter", "cell_param@+6"})",
-        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+4", "HadCalCell_PV@+1", "HadCalLayer_PV@+2"})",
-        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+2", "HadCalCell_PV@+1", "HadCalLayer_PV@+7"})",
-        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+3", "HadCalCell_PV@+1", "HadCalLayer_PV@+16"})",
-    };
-
-    EXPECT_VEC_EQ(expected_all_vol_inst, all_vol_inst);
-}
-
-//---------------------------------------------------------------------------//
-class TilecalPlugTest : public GeantGeoTest
-{
-    std::string geometry_basename() const final { return "tilecal-plug"; }
-};
-
-TEST_F(TilecalPlugTest, trace)
-{
-    {
-        SCOPED_TRACE("+z");
-        auto result = this->track({5.75, 0.01, -40}, {0, 0, 1});
-        static char const* const expected_volumes[] = {"Tile_ITCModule",
-                                                       "Tile_Plug1Module",
-                                                       "Tile_Absorber",
-                                                       "Tile_Plug1Module"};
-        EXPECT_VEC_EQ(expected_volumes, result.volumes);
-        static real_type const expected_distances[] = {22.9425, 0.115, 42, 37};
-        EXPECT_VEC_SOFT_EQ(expected_distances, result.distances);
-    }
-    {
-        SCOPED_TRACE("+z");
-        auto result = this->track({6.25, 0.01, -40}, {0, 0, 1});
-        static char const* const expected_volumes[]
-            = {"Tile_ITCModule", "Tile_Absorber", "Tile_Plug1Module"};
-        EXPECT_VEC_EQ(expected_volumes, result.volumes);
-        static real_type const expected_distances[] = {23.0575, 42, 37};
-        EXPECT_VEC_SOFT_EQ(expected_distances, result.distances);
-    }
-}
-
-//---------------------------------------------------------------------------//
-using TransformedBoxTest
-    = GenericGeoParameterizedTest<GeantGeoTest, TransformedBoxGeoTest>;
-
-TEST_F(TransformedBoxTest, accessors)
-{
-    this->impl().test_accessors();
-}
-
-TEST_F(TransformedBoxTest, trace)
-{
-    this->impl().test_trace();
-}
-
-//---------------------------------------------------------------------------//
-using CmseTest = GenericGeoParameterizedTest<GeantGeoTest, CmseGeoTest>;
-
-TEST_F(CmseTest, trace)
-{
-    this->impl().test_trace();
-}
-
-TEST_F(CmseTest, imager)
-{
-    SafetyImager write_image{this->geometry()};
-
-    ImageInput inp;
-    inp.lower_left = from_cm({-550, 0, -4000});
-    inp.upper_right = from_cm({550, 0, 2000});
-    inp.rightward = {0.0, 0.0, 1.0};
-    inp.vertical_pixels = 8;
-
-    std::string prefix = "g4";
-
-    write_image(ImageParams{inp}, prefix + "-cmse.jsonl");
-}
-
-//---------------------------------------------------------------------------//
-using CmsEeBackDeeTest
-    = GenericGeoParameterizedTest<GeantGeoTest, CmsEeBackDeeGeoTest>;
-
-TEST_F(CmsEeBackDeeTest, accessors)
-{
-    this->impl().test_accessors();
-}
-
-TEST_F(CmsEeBackDeeTest, trace)
-{
-    this->impl().test_trace();
-}
-
-//---------------------------------------------------------------------------//
-using ZnenvTest = GenericGeoParameterizedTest<GeantGeoTest, ZnenvGeoTest>;
-
-TEST_F(ZnenvTest, trace)
-{
-    this->impl().test_trace();
 }
 
 //---------------------------------------------------------------------------//
@@ -636,6 +440,194 @@ TEST_F(PincellTest, imager)
     inp.lower_left = from_cm({-12, 0, -12});
     inp.upper_right = from_cm({12, 0, 12});
     write_image(ImageParams{inp}, "g4-pincell-xz-mid.jsonl");
+}
+
+//---------------------------------------------------------------------------//
+using ReplicaTest = GenericGeoParameterizedTest<GeantGeoTest, ReplicaGeoTest>;
+
+TEST_F(ReplicaTest, trace)
+{
+    this->impl().test_trace();
+}
+
+TEST_F(ReplicaTest, volume_stack)
+{
+    this->impl().test_volume_stack();
+}
+
+TEST_F(ReplicaTest, level_strings)
+{
+    using R2 = Array<double, 2>;
+
+    auto const& geo_params = *this->geometry();
+    auto const& vol_inst = geo_params.volume_instances();
+
+    static R2 const points[] = {
+        {-435, 550},
+        {-460, 550},
+        {-400, 650},
+        {-450, 650},
+        {-450, 700},
+    };
+
+    std::vector<std::string> all_vol_inst;
+    for (R2 xz : points)
+    {
+        auto geo = this->make_geo_track_view({xz[0], 0.0, xz[1]}, {1, 0, 0});
+
+        auto level = geo.level();
+        CELER_ASSERT(level && level >= LevelId{0});
+        std::vector<VolumeInstanceId> inst_ids(level.get() + 1);
+        geo.volume_instance_id(make_span(inst_ids));
+        std::vector<std::string> names(inst_ids.size());
+        for (auto i : range(inst_ids.size()))
+        {
+            Label lab = vol_inst.at(inst_ids[i]);
+            if (auto phys_inst = geo_params.id_to_geant(inst_ids[i]))
+            {
+                if (phys_inst.replica)
+                {
+                    lab.ext += '+';
+                    lab.ext += std::to_string(phys_inst.replica.get());
+                }
+            }
+            names[i] = to_string(lab);
+        }
+        all_vol_inst.push_back(to_string(repr(names)));
+    }
+
+    static char const* const expected_all_vol_inst[] = {
+        R"({"world_PV", "fSecondArmPhys", "EMcalorimeter", "cell_param@+14"})",
+        R"({"world_PV", "fSecondArmPhys", "EMcalorimeter", "cell_param@+6"})",
+        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+4", "HadCalCell_PV@+1", "HadCalLayer_PV@+2"})",
+        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+2", "HadCalCell_PV@+1", "HadCalLayer_PV@+7"})",
+        R"({"world_PV", "fSecondArmPhys", "HadCalorimeter", "HadCalColumn_PV@+3", "HadCalCell_PV@+1", "HadCalLayer_PV@+16"})",
+    };
+
+    EXPECT_VEC_EQ(expected_all_vol_inst, all_vol_inst);
+}
+
+//---------------------------------------------------------------------------//
+class SolidsTest
+    : public GenericGeoParameterizedTest<GeantGeoTest, SolidsGeoTest>
+{
+    SpanStringView expected_log_levels() const final
+    {
+        if (geant4_version < Version{11})
+            return {};
+
+        static std::string_view const levels[] = {"error"};
+        return make_span(levels);
+    }
+};
+
+//---------------------------------------------------------------------------//
+TEST_F(SolidsTest, output)
+{
+    GeoParamsOutput out(this->geometry());
+    EXPECT_EQ("geometry", out.label());
+
+    if (CELERITAS_UNITS == CELERITAS_UNITS_CGS)
+    {
+        auto actual = to_string(out);
+        std::regex pattern(R"("",)");
+        actual = std::regex_replace(actual, pattern, "");
+        EXPECT_JSON_EQ(
+            R"json({"_category":"internal","_label":"geometry","bbox":[[-600.0,-300.0,-75.0],[600.0,300.0,75.0]],"max_depth":2,"supports_safety":true,"volumes":{"label":["box500","cone1","para1","sphere1","parabol1","trap1","trd1","trd2","trd3_refl@1","tube100","boolean1","polycone1","genPocone1","ellipsoid1","tetrah1","orb1","polyhedr1","hype1","elltube1","ellcone1","arb8b","arb8a","xtru1","World","trd3_refl@0"]}})json",
+            actual);
+    }
+}
+//---------------------------------------------------------------------------//
+
+TEST_F(SolidsTest, accessors)
+{
+    TestImpl(this).test_accessors();
+}
+
+//---------------------------------------------------------------------------//
+
+TEST_F(SolidsTest, trace)
+{
+    TestImpl(this).test_trace();
+}
+
+//---------------------------------------------------------------------------//
+
+TEST_F(SolidsTest, reflected_vol)
+{
+    auto geo = this->make_geo_track_view({-500, -125, 0}, {0, 1, 0});
+    EXPECT_EQ(25, geo.volume_id().unchecked_get() - this->volume_offset());
+    auto const& label = this->geometry()->volumes().at(geo.volume_id());
+    EXPECT_EQ("trd3_refl", label.name);
+    EXPECT_FALSE(ends_with(label.ext, "_refl"));
+}
+
+TEST_F(SolidsTest, DISABLED_imager)
+{
+    SafetyImager write_image{this->geometry()};
+
+    ImageInput inp;
+    inp.lower_left = from_cm({-550, -250, 5});
+    inp.upper_right = from_cm({550, 250, 5});
+    inp.rightward = {1.0, 0.0, 0.0};
+    inp.vertical_pixels = 8;
+
+    write_image(ImageParams{inp}, "g4-solids-xy-hi.jsonl");
+
+    inp.lower_left[2] = inp.upper_right[2] = from_cm(-5);
+    write_image(ImageParams{inp}, "g4-solids-xy-lo.jsonl");
+}
+
+//---------------------------------------------------------------------------//
+class TilecalPlugTest : public GeantGeoTest
+{
+    std::string geometry_basename() const final { return "tilecal-plug"; }
+};
+
+TEST_F(TilecalPlugTest, trace)
+{
+    {
+        SCOPED_TRACE("+z");
+        auto result = this->track({5.75, 0.01, -40}, {0, 0, 1});
+        static char const* const expected_volumes[] = {"Tile_ITCModule",
+                                                       "Tile_Plug1Module",
+                                                       "Tile_Absorber",
+                                                       "Tile_Plug1Module"};
+        EXPECT_VEC_EQ(expected_volumes, result.volumes);
+        static real_type const expected_distances[] = {22.9425, 0.115, 42, 37};
+        EXPECT_VEC_SOFT_EQ(expected_distances, result.distances);
+    }
+    {
+        SCOPED_TRACE("+z");
+        auto result = this->track({6.25, 0.01, -40}, {0, 0, 1});
+        static char const* const expected_volumes[]
+            = {"Tile_ITCModule", "Tile_Absorber", "Tile_Plug1Module"};
+        EXPECT_VEC_EQ(expected_volumes, result.volumes);
+        static real_type const expected_distances[] = {23.0575, 42, 37};
+        EXPECT_VEC_SOFT_EQ(expected_distances, result.distances);
+    }
+}
+
+//---------------------------------------------------------------------------//
+using TransformedBoxTest
+    = GenericGeoParameterizedTest<GeantGeoTest, TransformedBoxGeoTest>;
+
+TEST_F(TransformedBoxTest, accessors)
+{
+    this->impl().test_accessors();
+}
+
+TEST_F(TransformedBoxTest, trace)
+{
+    this->impl().test_trace();
+}
+
+//---------------------------------------------------------------------------//
+using ZnenvTest = GenericGeoParameterizedTest<GeantGeoTest, ZnenvGeoTest>;
+
+TEST_F(ZnenvTest, trace)
+{
+    this->impl().test_trace();
 }
 
 //---------------------------------------------------------------------------//
